Moved project state mixin to project, because the states don't read from the mixin for whatever reason
